# [WIP][ENH,MAINT] Accelerate HPI fit and add freq ordering to plugin 