… examples/ex_hpiFit/main.cpp
… libraries/disp/viewers/formfiles/hpisettingsview.ui
… libraries/inverse/hpiFit/hpifit.h
… libraries/inverse/hpiFit/hpifitdata.cpp
… libraries/inverse/hpiFit/hpifitdata.h
… libraries/rtprocessing/rthpis.cpp
… libraries/rtprocessing/rthpis.h
… testframes/test_hpiFit/test_hpiFit.cpp