This PR adds the frequency ordering to the hpi plugin in mne-scan and improves the performance of the hpi fitting. 

**frequency ordering:**
- add a push button to control panel and implement it into the hpi-plugin

**performance improvement:**
I was able to improve the performance of the hpi fit from **~250 ms** per fit to **~65 ms** per fit on my laptop. I got this improvement by:

- move calculation of everything that only has to be calculated once to the constructor (~225 ms)
- optimize parameters of minimization algorithm fminsearch (~112 ms)
- replace qlist in Sensor Struct with Eigen Matrices (~65 ms)

The result is shown in the following figure, just to give an impression. The values are calculated on one .fif file and all in all ~360 fits were performed. The figure shows the result of each improvement described above.
![Figure_1](https://user-images.githubusercontent.com/40984141/77914568-6a29b300-7296-11ea-8cce-8d5e3496b9ab.png)

**Everything else**
- apply changes in the use of hpifit class everywhere where needed
- apply pedantic naming conventions