@RDoerfel 2020-03-30
[WIP] because I have to remove file path changes from ex_hpiFit once I'm sure there will be no further changes to make